[wait-merge] Use 'gdm' to delete merged branch
(This will also delete other branches that don't differ from the primary
branch.)

Motivation: the `git branch -d` command was failing (silently, due to us
suppressing the output) with a warning about the branch not being fully
merged.

https://stackoverflow.com/questions/7548926/git-error-the-branch-x-is-not-fully-merged

One option would be to use `-D` rather than `-d` to force the deletion,
but I guess that I feel more comfortable using `gdm` to only delete the
branch if it really has been merged.

Since I think that this will fix the issue that was causing us not to
reach the end of the script, I'm also removing the debugging changes
that I had made to this file.