Move Runger.config.log_capybara? into patched Capybara method
This makes it possible to toggle it on and off at runtime. Before, if we
start up with `log_capybara?` false, then we aren't able to turn it on
and have that take effect while the process is still running (e.g. when
using `gal`). This change allows us to toggle it on/off without needing
to restart the `gal` process.