Don't monkeypatch Rollbar if SKIP_ROLLBAR_MONKEYPATCH env var is set
This can be useful when using `gal -g sidekiq`, for example. When there
is an error, our Rollbar monkeypatch prints out a big stack trace, so I
think need to scroll up to see the actual error message, which is
annoying and inefficient.

Now, by doing something like `SKIP_ROLLBAR_MONKEYPATCH=1 gal -g
sidekiq`, instead, the Rollbar monkeypatch will not be applied, and so
it won't print error backtraces.